refactor: move Status.ping() into Gateway.status()
