+++ src/models/query_contract.py
@@ -5,7 +5,7 @@ import traceback
 import logging
 
 from dependencies import title
-from models.query import Query
+from lib.query import Query
 from init import DB
 
 from config import URL
